Struct wasmparser::collections::set::Set

pub struct Set<T> { /* private fields */ }

A default set of values.

Provides an API compatible with both HashSet and BTreeSet.

Implementations

impl<T> Set<T>

pub fn clear(&mut self)

Clears the Set, removing all elements.

pub fn retain<F>(&mut self, f: F)
where T: Ord, F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements e for which f(&e) returns false. The elements are visited in unsorted (and unspecified) order.

pub fn len(&self) -> usize

Returns the number of elements in the Set.

pub fn is_empty(&self) -> bool

Returns true if the Set contains no elements.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator that yields the items in the Set.

impl<T> Set<T>
where T: Eq + Hash + Ord,

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the Set.

pub fn contains<Q>(&self, value: &Q) -> bool
where T: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Returns true if the Set contains an element equal to the value.

pub fn get<Q>(&self, value: &Q) -> Option<&T>
where T: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

Returns a reference to the element in the Set, if any, that is equal to the value.

pub fn insert(&mut self, value: T) -> bool

Adds value to the Set.

Returns whether the value was newly inserted:

Returns true if the set did not previously contain an equal value.

Returns false otherwise and the entry is not updated.

pub fn remove<Q>(&mut self, value: &Q) -> bool
where T: Borrow<Q>, Q: ?Sized + Hash + Eq + Ord,

If the set contains an element equal to the value, removes it from the Set and drops it.

Returns true if such an element was present, otherwise false.

pub fn take<Q>(&mut self, value: &Q) -> Option<T>
where T: Borrow<Q>, Q: ?Sized + Hash + Ord,

Removes and returns the element in the Set, if any, that is equal to the value.

The value may be any borrowed form of the setโs element type, but the ordering on the borrowed form must match the ordering on the element type.

pub fn replace(&mut self, value: T) -> Option<T>

Adds a value to the Set, replacing the existing value, if any, that is equal to the given one. Returns the replaced value.

pub fn is_disjoint(&self, other: &Self) -> bool

Returns true if self has no elements in common with other. This is equivalent to checking for an empty intersection.

pub fn is_subset(&self, other: &Self) -> bool

Returns true if the Set is a subset of another, i.e., other contains at least all the values in self.

pub fn is_superset(&self, other: &Self) -> bool

Returns true if the Set is a superset of another, i.e., self contains at least all the values in other.

pub fn difference<'a>(&'a self, other: &'a Self) -> Difference<'a, T>

Visits the values representing the difference, i.e., the values that are in self but not in other.

pub fn symmetric_difference<'a>( &'a self, other: &'a Self, ) -> SymmetricDifference<'a, T>

Visits the values representing the symmetric difference, i.e., the values that are in self or in other but not in both.

pub fn intersection<'a>(&'a self, other: &'a Self) -> Intersection<'a, T>

Visits the values representing the intersection, i.e., the values that are both in self and other.

When an equal element is present in self and other then the resulting Intersection may yield references to one or the other. This can be relevant if T contains fields which are not compared by its Eq implementation, and may hold different value between the two equal copies of T in the two sets.

pub fn union<'a>(&'a self, other: &'a Self) -> Union<'a, T>

Visits the values representing the union, i.e., all the values in self or other, without duplicates.