Enum wasmparser::Name

pub enum Name<'a> {

    Module {
        name: &'a str,
        name_range: Range<usize>,
    },
    Function(NameMap<'a>),
    Local(IndirectNameMap<'a>),
    Label(IndirectNameMap<'a>),
    Type(NameMap<'a>),
    Table(NameMap<'a>),
    Memory(NameMap<'a>),
    Global(NameMap<'a>),
    Element(NameMap<'a>),
    Data(NameMap<'a>),
    Field(IndirectNameMap<'a>),
    Tag(NameMap<'a>),
    Unknown {
        ty: u8,
        data: &'a [u8],
        range: Range<usize>,
    },

}

Represents a name read from the names custom section.

Variants

Module

The name is for the module.

Fields

name: &'a str

The specified name.

name_range: Range<usize>

The byte range that name occupies in the original binary.

Function(NameMap<'a>)

The name is for the functions.

Local(IndirectNameMap<'a>)

The name is for the function locals.

Label(IndirectNameMap<'a>)

The name is for the function labels.

Type(NameMap<'a>)

The name is for the types.

Table(NameMap<'a>)

The name is for the tables.

Memory(NameMap<'a>)

The name is for the memories.

Global(NameMap<'a>)

The name is for the globals.

Element(NameMap<'a>)

The name is for the element segments.

Data(NameMap<'a>)

The name is for the data segments.

Field(IndirectNameMap<'a>)

The name is for fields.

Tag(NameMap<'a>)

The name is for tags.

Unknown

An unknown name subsection.

Fields

ty: u8

The identifier for this subsection.

data: &'a [u8]

The contents of this subsection.

range: Range<usize>

The range of bytes, relative to the start of the original data stream, that the contents of this subsection reside in.