wasmparser
Module names

Definitions of name-related helpers and newtypes, primarily for the component model.

Structs

ComponentName

An import or export name in the component model which is backed by T, which defaults to String.

DependencyName

A dependency on an implementation either as locked-dep=... or unlocked-dep=...

HashName

A dependency on an implementation either as integrity=....

InterfaceName

An interface name, stored as a:b/c@1.2.3

KebabStr

Represents a kebab string slice used in validation.

KebabString

Represents an owned kebab string for validation.

ResourceFunc

A resource name and its function, stored as a.b.

UrlName

A dependency on an implementation either as url=...

Enums

ComponentNameKind

Created via ComponentName::kind and classifies a name.