Struct wasmparser::RefType

pub struct RefType(_);

A reference type.

The reference types proposal first introduced externref and funcref.

The function references proposal introduced typed function references.

The GC proposal introduces heap types: any, eq, i31, struct, array, nofunc, noextern, none.

Implementations

impl RefType

pub const FUNCREF: Self = _

A nullable untyped function reference aka (ref null func) aka funcref aka anyfunc.

pub const EXTERNREF: Self = _

A nullable reference to an extern object aka (ref null extern) aka externref.

pub const FUNC: Self = _

A non-nullable untyped function reference aka (ref func).

pub const EXTERN: Self = _

A non-nullable reference to an extern object aka (ref extern).

pub const ANY: Self = _

A non-nullable reference to any object aka (ref any).

pub const NONE: Self = _

A non-nullable reference to no object aka (ref none).

pub const NOEXTERN: Self = _

A non-nullable reference to a noextern object aka (ref noextern).

pub const NOFUNC: Self = _

A non-nullable reference to a nofunc object aka (ref nofunc).

pub const EQ: Self = _

A non-nullable reference to an eq object aka (ref eq).

pub const STRUCT: Self = _

A non-nullable reference to a struct aka (ref struct).

pub const ARRAY: Self = _

A non-nullable reference to an array aka (ref array).

pub const I31: Self = _

A non-nullable reference to an i31 object aka (ref i31).

pub const fn indexed_func(nullable: bool, index: u32) -> Option<Self>

Create a reference to a typed function with the type at the given index.

Returns None when the type index is beyond this crate’s implementation limits and therefore is not representable.

pub const fn indexed_array(nullable: bool, index: u32) -> Option<Self>

Create a reference to an array with the type at the given index.

Returns None when the type index is beyond this crate’s implementation limits and therefore is not representable.

pub const fn indexed_struct(nullable: bool, index: u32) -> Option<Self>

Create a reference to a struct with the type at the given index.

Returns None when the type index is beyond this crate’s implementation limits and therefore is not representable.

pub const fn new(nullable: bool, heap_type: HeapType) -> Option<Self>

Create a new RefType.

Returns None when the heap type’s type index (if any) is beyond this crate’s implementation limits and therfore is not representable.

pub const fn is_typed_func_ref(&self) -> bool

Is this a reference to a typed function?

pub const fn is_indexed_type_ref(&self) -> bool

Is this a reference to an indexed type?

pub const fn type_index(&self) -> Option<u32>

If this is a reference to a typed function, get its type index.

pub const fn is_func_ref(&self) -> bool

Is this an untyped function reference aka (ref null func) aka funcref aka anyfunc?

pub const fn is_extern_ref(&self) -> bool

Is this a (ref null extern) aka externref?

pub const fn is_nullable(&self) -> bool

Is this ref type nullable?

pub const fn as_non_null(&self) -> Self

Get the non-nullable version of this ref type.

pub const fn nullable(&self) -> Self

Get the non-nullable version of this ref type.

pub fn heap_type(&self) -> HeapType

Get the heap type that this is a reference to.