wasmparser

Struct WasmFeatures

pub struct WasmFeatures { /* private fields */ }

Enabled WebAssembly proposals and features.

This is the disabled zero-size version of this structure because the features feature was disabled at compile time of this crate.

Implementations

impl WasmFeatures

pub fn mutable_global(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn saturating_float_to_int(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn sign_extension(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn reference_types(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn multi_value(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn bulk_memory(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn simd(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn relaxed_simd(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn threads(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn shared_everything_threads(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn tail_call(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn floats(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn multi_memory(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn exceptions(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn memory64(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn extended_const(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn component_model(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn function_references(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn memory_control(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn gc(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn custom_page_sizes(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn legacy_exceptions(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn gc_types(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn stack_switching(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn wide_arithmetic(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_values(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_nested_names(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_async(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_async_stackful(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_async_builtins(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_error_context(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_fixed_size_list(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn cm_gc(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn call_indirect_overlong(&self) -> bool

Returns whether this feature is enabled in this feature set.

pub fn bulk_memory_opt(&self) -> bool

Returns whether this feature is enabled in this feature set.