wasmparser

Trait FrameStack

pub trait FrameStack {
    // Required method
    fn current_frame(&self) -> Option<FrameKind>;
}

A trait representing the stack of frames within a function.

The BinaryReader::visit_operator and OperatorsReaders type use information about the current frame kind to enforce the syntactic requirements of the binary format.

Required Methods

fn current_frame(&self) -> Option<FrameKind>

The current frame kind.