Trait wasmparser::SectionWithLimitedItems

pub trait SectionWithLimitedItems: SectionReader {
    // Required method
    fn get_count(&self) -> u32;

    // Provided method
    fn into_iter_with_offsets(self) -> IntoIterWithOffsets<Self> 
       where Self: Sized { ... }
}

Implemented by sections with a limited number of items.

Required Methods

fn get_count(&self) -> u32

Gets the count of the items in the section.

Provided Methods

fn into_iter_with_offsets(self) -> IntoIterWithOffsets<Self> where Self: Sized,

Returns an iterator over the items within this section where the offset in the original section is provided with the item.