Trait wasmparser::Subsection

pub trait Subsection<'a>: Sized {
    fn from_reader(id: u8, reader: BinaryReader<'a>) -> Result<Self>;
}

A trait implemented for subsections of another outer section.

This is currently only used for subsections within custom sections, such as the name section of core wasm.

This is used in conjunction with Subsections.

Required Methods

fn from_reader(id: u8, reader: BinaryReader<'a>) -> Result<Self>

Converts the section identifier provided with the section contents into a typed section