Trait wasmparser::WasmFuncType

pub trait WasmFuncType {
    type Type: WasmType;
    fn len_inputs(&self) -> usize;

    fn len_outputs(&self) -> usize;

    fn input_at(&self, at: u32) -> Option<&Self::Type>;

    fn output_at(&self, at: u32) -> Option<&Self::Type>;
}

Types that qualify as Wasm function types for validation purposes.

Associated Types

type Type: WasmType

A type that is comparable with Wasm types.

Loading content...
Required methods

fn len_inputs(&self) -> usize

Returns the number of input types.

fn len_outputs(&self) -> usize

Returns the number of output types.

fn input_at(&self, at: u32) -> Option<&Self::Type>

Returns the type at given index if any.

Note

The returned type may be wrapped by the user crate and thus the actually returned type only has to be comparable to a Wasm type.

fn output_at(&self, at: u32) -> Option<&Self::Type>

Returns the type at given index if any.

Note

The returned type may be wrapped by the user crate and thus the actually returned type only has to be comparable to a Wasm type.

Loading content...