Trait wasmparser::WasmMemoryType

pub trait WasmMemoryType {
    fn is_shared(&self) -> bool;

    fn initial_limit(&self) -> u32;

    fn maximum_limit(&self) -> Option<u32>;
}

Types that qualify as Wasm memory types for validation purposes.

Required methods

fn is_shared(&self) -> bool

Returns true if the linear memory is shared.

fn initial_limit(&self) -> u32

Returns the initial limit of the linear memory.

fn maximum_limit(&self) -> Option<u32>

Returns the maximum limit of the linear memory if any.

Loading content...