Type Alias wasmparser::CoreTypeSectionReader

pub type CoreTypeSectionReader<'a> = SectionLimited<'a, CoreType<'a>>;

A reader for the core type section of a WebAssembly component.

Examples

use wasmparser::CoreTypeSectionReader;
let mut reader = CoreTypeSectionReader::new(data, 0).unwrap();
for ty in reader {
    println!("Type {:?}", ty.expect("type"));
}

Aliased Type

struct CoreTypeSectionReader<'a> { /* private fields */ }

Implementations

impl<'a, T> SectionLimited<'a, T>

pub fn new(data: &'a [u8], offset: usize) -> Result<Self>

Creates a new section reader from the provided contents.

The data provided here is the data of the section itself that will be parsed. The offset argument is the byte offset, in the original wasm binary, that the section was found. The offset argument is used for error reporting.

Errors

Returns an error if a 32-bit count couldn’t be read from the data.

pub fn count(&self) -> u32

Returns the count of total items within this section.

pub fn original_position(&self) -> usize

Returns whether the original byte offset of this section.

pub fn range(&self) -> Range<usize>

Returns the range, as byte offsets, of this section within the original wasm binary.

pub fn into_iter_with_offsets(self) -> SectionLimitedIntoIterWithOffsets<'a, T> where T: FromReader<'a>,

Returns an iterator which yields not only each item in this section but additionally the offset of each item within the section.