Struct wasmparser::types::SubtypeArena

pub struct SubtypeArena<'a> { /* private fields */ }

A helper typed used purely during subtyping as part of SubtypeCx.

This takes a types list as input which is the “base” of the ids that can be indexed through this arena. All future types pushed into this, if any, are stored in self.list.

This is intended to have arena-like behavior where everything pushed onto self.list is thrown away after a subtyping computation is performed. All new types pushed into this arena are purely temporary.