wasmparser::types

Trait TypeData

pub trait TypeData: Debug {
    type Id: TypeIdentifier<Data = Self>;

    const IS_CORE_SUB_TYPE: bool;
}

A trait shared by all types within a Types.

This is the data that can be retrieved by indexing with the associated TypeIdentifier.

Required Associated Constants

const IS_CORE_SUB_TYPE: bool

Is this type a core sub type (or rec group of sub types)?

Required Associated Types

type Id: TypeIdentifier<Data = Self>

The identifier for this type data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.