wasmparser::types

Trait TypeIdentifier

pub trait TypeIdentifier:
    Debug
    + Copy
    + Eq
    + Sized
    + 'static {
    type Data: TypeData<Id = Self>;
}

A trait shared by all type identifiers.

Any id that can be used to get a type from a Types.

Required Associated Types

type Data: TypeData<Id = Self>

The data pointed to by this type of id.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.