wasmparser

Module collections

Type definitions for maps and sets used by the wasmparser crate.

This module contains type definitions for Map, Set, IndexMap, and IndexSet. These types are thin-wrappers around either hash-map based or B-tree-map based data structures. Users can strictly use the btree-map based variants by enabling the no-hash-maps crate feature.

Map: Either backed by [hashbrown::HashMap] or Rust’s BTreeMap.

Set: Either backed by [hashbrown::HashSet] or Rust’s BTreeSet.

IndexMap: Either backed by [indexmap::IndexMap] or a custom implementation based on Rust’s BTreeMap.

IndexSet: Either backed by [indexmap::IndexSet] or a custom implementation based on Rust’s BTreeMap.

For the hash-map based type definitions the hash algorithm type parameter is fixed.

Modules

index_map

Type definitions for an ordered map.

index_set

Type definitions for an ordered set.

map

Type definitions for a default map.

set

Type definitions for a default set.

Structs

IndexMap

A hash table where the iteration order of the key-value pairs is independent of the hash values of the keys.

IndexSet

A default set of values.

Map

A default key-value mapping.

Set

A default set of values.