Trait wasmparser::FromReader

pub trait FromReader<'a>: Sized {
    fn from_reader(reader: &mut BinaryReader<'a>) -> Result<Self>;
}

A trait implemented for items that can be decoded directly from a BinaryReader, or that which can be parsed from the WebAssembly binary format.

Note that this is also accessible as a BinaryReader::read method.

Required Methods

fn from_reader(reader: &mut BinaryReader<'a>) -> Result<Self>

Attempts to read Self from the provided binary reader, returning an error if it is unable to do so.