wasmparser

Type Alias TypeSectionReader

pub type TypeSectionReader<'a> = SectionLimited<'a, RecGroup>;

A reader for the type section of a WebAssembly module.

Aliased Type

struct TypeSectionReader<'a> { /* private fields */ }

Implementations

impl<'a> TypeSectionReader<'a>

pub fn into_iter_err_on_gc_types( self, ) -> impl Iterator<Item = Result<FuncType>> + 'a

Returns an iterator over this type section which will only yield function types and any usage of GC types from the GC proposal will be translated into an error.